finished friends side container display
