+++ src/styles/main.css
@@ -35,8 +35,8 @@
 .friend-card {
   border: solid black 1px;
   padding: 2px;
-  background-color: grey;
-  margin-bottom: 1px;
+  background-color: #6c757d;
+  margin-bottom: 2px;
   border-radius: .25rem;
 }
